+++ packages/template-retail-react-app/config/default.js
@@ -79,7 +79,10 @@ module.exports = {
             }
         },
         storeLocatorEnabled: true,
-        multishipEnabled: true
+        multishipEnabled: true,
+        googleCloudAPI: {
+            apiKey: process.env.GOOGLE_CLOUD_API_KEY
+        }
     },
     envBasePath: '/',
     externals: [],